fix: 상품 등록, 수정, 삭제

+++ src/main/java/kr/casealot/shop/domain/product/service/ProductService.java
@@ -1,5 +1,8 @@
 package kr.casealot.shop.domain.product.service;
 
+import io.jsonwebtoken.Claims;
+import kr.casealot.shop.domain.customer.entity.Customer;
+import kr.casealot.shop.domain.customer.repository.CustomerRepository;
 import kr.casealot.shop.domain.product.dto.ProductDTO;
 import kr.casealot.shop.domain.product.dto.ProductReqDTO;
 import kr.casealot.shop.domain.product.dto.ProductResDTO;
@@ -8,6 +11,9 @@ import kr.casealot.shop.domain.product.entity.Product;
 import kr.casealot.shop.domain.product.repository.ProductRepository;
 import kr.casealot.shop.domain.product.support.ProductSpecification;
 import kr.casealot.shop.domain.qna.entity.Qna;
+import kr.casealot.shop.global.oauth.token.AuthToken;
+import kr.casealot.shop.global.oauth.token.AuthTokenProvider;
+import kr.casealot.shop.global.util.HeaderUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.data.domain.Page;
@@ -15,12 +21,15 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
+import static kr.casealot.shop.global.oauth.entity.RoleType.ADMIN;
 import static org.springframework.data.crossstore.ChangeSetPersister.*;
 
 @Service
@@ -28,9 +37,19 @@ import static org.springframework.data.crossstore.ChangeSetPersister.*;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final CustomerRepository customerRepository;
+    private final AuthTokenProvider authTokenProvider;
 
     @Transactional
-    public void createProduct(ProductDTO productDTO) {
+    public void createProduct(ProductDTO productDTO, HttpServletRequest request) {
+
+        String token = HeaderUtil.getAccessToken(request);
+        AuthToken authToken = authTokenProvider.convertAuthToken(token);
+        Claims claims = authToken.getTokenClaims();
+        String customerId = claims.getSubject();
+
+        Customer customer = customerRepository.findById(customerId);
+
         Product product = Product.builder()
                 .name(productDTO.getName())
                 .content(productDTO.getContent())
@@ -45,6 +64,7 @@ public class ProductService {
                 .color(productDTO.getColor())
                 .season(productDTO.getSeason())
                 .type(productDTO.getType())
+                .customer(customer)
                 .build();
 
         productRepository.save(product);
@@ -52,11 +72,21 @@ public class ProductService {
 
 
     @Transactional
-    public void updateProduct(Long productId, ProductDTO productDTO) {
-        Product product = productRepository.findById(productId)
-                .orElseThrow();
+    public void updateProduct(Long productId, ProductDTO productDTO,
+                              HttpServletRequest request) throws NotFoundException {
 
-        product.setName(product.getName());
+        Product product = productRepository.findById(productId)
+                .orElseThrow(NotFoundException::new);
+
+        String customerId = findCustomerId(request);
+
+        boolean isAdmin = checkAdminRole(customerId);
+
+        if (!isAdmin) {
+            throw new AccessDeniedException("You are not authorized to update this Product.");
+        }
+
+        product.setName(productDTO.getName());
         product.setContent(productDTO.getContent());
         product.setImg_B(productDTO.getImg_B());
         product.setImg_M(productDTO.getImg_M());
@@ -70,11 +100,19 @@ public class ProductService {
         productRepository.save(product);
     }
 
-
-    public void deleteProduct(Long productId) throws NotFoundException {
+    @Transactional
+    public void deleteProduct(Long productId, HttpServletRequest request) throws NotFoundException {
 
         Product product = productRepository.findById(productId)
                 .orElseThrow(NotFoundException::new);
+
+        String customerId = findCustomerId(request);
+
+        boolean isAdmin = checkAdminRole(customerId);
+
+        if (!isAdmin) {
+            throw new AccessDeniedException("You are not authorized to delete this product.");
+        }
 
         productRepository.delete(product);
     }
@@ -115,6 +153,18 @@ public class ProductService {
         savedProduct.setViews(savedProduct.getViews() + 1);
         productRepository.save(savedProduct);
         return savedProduct;
+    }
+
+    private String findCustomerId(HttpServletRequest request) {
+        String token = HeaderUtil.getAccessToken(request);
+        AuthToken authToken = authTokenProvider.convertAuthToken(token);
+        Claims claims = authToken.getTokenClaims();
+        return claims.getSubject();
+    }
+
+    private boolean checkAdminRole(String customerId) {
+        Customer customer = customerRepository.findById(customerId);
+        return customer.getRoleType() == ADMIN;
     }
 
 }
